fix error

+++ sayurbox/src/main/java/com/sayurbox/sayurbox/models/Product.java
@@ -16,7 +16,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int productId;
+    private Long productId;
 
     @Column(nullable = false)
     private String name;
@@ -44,11 +44,11 @@ public class Product {
     }
 
     // Getters and Setters
-    public int getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
-    public void setProductId(int productId) {
+    public void setProductId(Long productId) {
         this.productId = productId;
     }
